feat: Implement core chess engine components including board, move ordering, search, evaluation, and UCI protocol.

+++ include/move.hpp
@@ -135,10 +135,10 @@ inline Move string_to_move(const std::string& str) {
 
 struct ScoredMove {
     Move move;
-    int score;
+    int64_t score;
 
     constexpr ScoredMove() : move(MOVE_NONE), score(0) {}
-    constexpr ScoredMove(Move m, int s) : move(m), score(s) {}
+    constexpr ScoredMove(Move m, int64_t s) : move(m), score(s) {}
 
     constexpr bool operator<(const ScoredMove& other) const {
         return score > other.score;  // Higher score = better move
@@ -161,7 +161,7 @@ public:
         }
     }
 
-    void add(Move m, int score) {
+    void add(Move m, int64_t score) {
         if (count < MAX_MOVES) {
             moves[count++] = ScoredMove(m, score);
         }
